Address review: init _target_key_to_idx in __init__, move tests
- Assign _target_key_to_idx in TinyLoraModel.__init__ instead of lazily
  in _create_and_replace, since super().__init__ already triggers the
  first injection cycle that reads it.
- Move the two new regression tests from the standalone tests/test_tinylora.py
  into tests/test_initialization.py as TestTinyLoraInitialization, matching
  this repo's test-location convention.
